distance is now an integer
git-svn-id: http://redbus.googlecode.com/svn/trunk@194 047a1783-e752-e205-4d80-9245024c4953

+++ src/org/redbus/BusTimesActivity.java
@@ -401,7 +401,7 @@ public class BusTimesActivity extends ListActivity implements BusDataResponseLis
 					i.putExtra("StopCode", stopCode);
 					i.putExtra("StopName", stopName);
 					i.putExtra("Location", location);
-					i.putExtra("Distance", (double) proximityAlarmDistances[distanceSpinner.getSelectedItemPosition()]);
+					i.putExtra("Distance", proximityAlarmDistances[distanceSpinner.getSelectedItemPosition()]);
 					i.putExtra("StartTime", System.currentTimeMillis());
 					PendingIntent pi = PendingIntent.getBroadcast(BusTimesActivity.this, 0, i, PendingIntent.FLAG_UPDATE_CURRENT);
